deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
plotly
Pillow
catboost
feature-engine
scikit-learn==1.2.2